Move `fe/` to `timemachine/fe`
