+++ examples/estimator_variance.py
@@ -27,8 +27,8 @@ import os
 import numpy as np
 from rdkit import Chem
 
-from fe import free_energy, topology
-from fe.free_energy import construct_lambda_schedule
+from timemachine.fe import topology, free_energy
+from timemachine.fe.free_energy import construct_lambda_schedule
 from ff import Forcefield
 from ff.handlers.deserialize import deserialize_handlers
 from md import builders, minimizer
