+++ examples/rbfe_single.py
@@ -8,9 +8,9 @@ import numpy as np
 import jax
 from jax import numpy as jnp
 
-from fe.free_energy import construct_lambda_schedule
-from fe.utils import convert_uIC50_to_kJ_per_mole
-from fe import model
+from timemachine.fe.free_energy import construct_lambda_schedule
+from timemachine.fe.utils import convert_uIC50_to_kJ_per_mole
+from timemachine.fe import model
 from md import builders
 
 from testsystems.relative import hif2a_ligand_pair
