+++ md/minimizer.py
@@ -1,13 +1,12 @@
 import numpy as np
 
-from fe import topology
+from timemachine.fe import topology, model_utils
 
 from timemachine.lib import LangevinIntegrator, MonteCarloBarostat, custom_ops
 
 from ff.handlers import openmm_deserializer
 from ff import Forcefield
-from fe import model_utils
-from fe.utils import get_romol_conf
+from timemachine.fe.utils import get_romol_conf
 from md.barostat.utils import get_group_indices, get_bond_list
 
 from rdkit import Chem
