+++ tests/dual_topology.py
@@ -13,10 +13,10 @@ from timemachine.lib import LangevinIntegrator
 from ff.handlers import openmm_deserializer
 from ff.handlers.deserialize import deserialize_handlers
 
-from fe import pdb_writer
-from fe.utils import get_romol_conf
-from fe import rbfe
-from md import Recipe
+from timemachine.fe import pdb_writer
+from timemachine.fe.utils import get_romol_conf
+from timemachine.fe import rbfe  # unresolved
+from md import Recipe  # unresolved
 from md import builders
 
 from multiprocessing import Pool
