+++ tests/test_estimator.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 
-from fe import estimator_abfe
+from timemachine.fe import estimator_abfe
 from timemachine.lib import LangevinIntegrator, potentials, MonteCarloBarostat
 from parallel.client import CUDAPoolClient
 from md.barostat.utils import get_bond_list, get_group_indices
