+++ tests/test_frame_filters.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
-from fe.estimator_abfe import SimulationResult
-from fe.frames import endpoint_frames_only, all_frames, no_frames
+from timemachine.fe.estimator_abfe import SimulationResult
+from timemachine.fe.frames import endpoint_frames_only, all_frames, no_frames
 
 
 def make_dummy_sim_result(val):
