+++ tests/test_nblist.py
@@ -3,7 +3,7 @@ import numpy as np
 
 from timemachine.lib import custom_ops
 from md.builders import build_water_system
-from fe.utils import get_romol_conf
+from timemachine.fe.utils import get_romol_conf
 from hilbertcurve.hilbertcurve import HilbertCurve
 from testsystems.relative import hif2a_ligand_pair
