+++ tests/test_standard_state.py
@@ -1,8 +1,7 @@
 import numpy as np
 import functools
 
-from simtk import unit
-from fe import standard_state
+from timemachine.fe import standard_state
 import rmsd
 
 import scipy.integrate
