+++ testsystems/relative.py
@@ -3,7 +3,7 @@
 import numpy as np
 from rdkit import Chem
 
-from fe import free_energy, topology
+from timemachine.fe import topology, free_energy
 from ff import Forcefield
 from ff.handlers.deserialize import deserialize_handlers
