+++ timemachine/fe/model.py
@@ -8,8 +8,8 @@ from rdkit import Chem
 
 from md import minimizer
 from timemachine.lib import LangevinIntegrator, MonteCarloBarostat
-from fe import free_energy, topology, estimator
-from fe.model_utils import apply_hmr
+from timemachine.fe import topology, estimator, free_energy
+from timemachine.fe.model_utils import apply_hmr
 from ff import Forcefield
 from md.barostat.utils import get_bond_list, get_group_indices
