+++ timemachine/fe/model_rabfe.py
@@ -7,8 +7,9 @@ from rdkit import Chem
 
 from timemachine.lib import potentials, LangevinIntegrator, MonteCarloBarostat
 from timemachine import constants
-from fe.frames import endpoint_frames_only
-from fe import free_energy_rabfe, topology, estimator_abfe, model_utils
+from timemachine.fe.frames import endpoint_frames_only
+from timemachine.fe import estimator_abfe, model_utils
+from timemachine.fe import topology, free_energy_rabfe
 from ff import Forcefield
 
 from parallel.client import AbstractClient, _MockFuture
